// src/state/action-types.ts
export enum GAME_ACTION_TYPES {
    SET_MINES_COUNTER = 'SET_MINES_COUNTER',
    SET_TIME_COUNTER = 'SET_TIME_COUNTER',
    SET_SMILEY_BUTTON = 'SET_SMILEY_BUTTON',
    ON_MOUSE_DOWN = 'ON_MOUSE_DOWN',
    ON_MOUSE_UP = 'ON_MOUSE_UP',
    ON_MOUSE_ENTER = 'ON_MOUSE_ENTER',
    CLOCK_TICK = 'CLOCK_TICK',
    CELL_LEFT_CLICK = 'CELL_LEFT_CLICK',
    CELL_RIGHT_CLICK = 'CELL_RIGHT_CLICK',
    CELL_BOTH_CLICK = 'CELL_BOTH_CLICK',
    SMILEY_CLICK = 'SMILEY_CLICK',
}